Allow cross-OS installation of wasi-sdk
This alters the action slightly to use a Python script to download and
extract the release artifacts for a given wasi-sdk version. Python,
IMHO, even with different versions across different runner OSes, is
still a more reliable way to do cross-platform things in GitHub actions.
The script is also factored to allow for running and testing locally.

This also improves the testing of the action with some unit tests. The
CI workflow is refactored a bit as well: now we can run all of the good
tests that were present previously for each possible combination of a
matrix.